updated to .net 4.8 (same as FW)
