+++ src/IcuEC/IcuEC 2010.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -10,7 +10,7 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>IcuEC</RootNamespace>
     <AssemblyName>IcuEC</AssemblyName>
-    <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <FileUpgradeFlags>
     </FileUpgradeFlags>
@@ -98,9 +98,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\icu4net\ICU4NETExtension\IcuBreakIteratorEncConverter.cs">
-      <Link>IcuBreakIteratorEncConverter.cs</Link>
-    </Compile>
+    <Compile Include="IcuBreakIteratorEncConverter.cs" />
     <Compile Include="IcuBreakIteratorAutoConfigDialog.cs">
       <SubType>Form</SubType>
     </Compile>
